requirements.txt
anyjson (0.3.3)
Flask (0.10.1)
Flask-SQLAlchemy (2.0)
itsdangerous (0.24)
Jinja2 (2.7.3)
jsonschema (2.4.0)
MarkupSafe (0.23)
pip (1.5.4)
pycrypto (2.6.1)
Pygments (2.0.1)
PyJWT (0.3.0)
PyYAML (3.11)
SQLAlchemy (0.9.8)
urwid (1.3.0)
Werkzeug (0.9.6)
